backend-service/requirements.txt
certifi==2021.5.30
Django==3.2.8
djangorestframework==3.12.4
python-dotenv==0.19.1
